requirements.txt
googlesearch-python==1.3.0